refactor: move message box interfaces to api package Signed-off-by: Florent Benoit <fbenoit@redhat.com>
